# Speed up the Base62.compatible? check by 10x