--- a/.gitignore
+++ b/.gitignore
@@ -8,4 +8,5 @@
 /spec/examples.txt
 /spec/reports/
 /tmp/
+*.jsonld
 Gemfile.lock