This change _actually_ applies the optimization I intended to in 2021.

<details>
<summary>Results of benchmark</summary>

```
Warming up --------------------------------------
            include?    22.058k i/100ms
       Regexp#match?   221.543k i/100ms
       #match?(char)    18.529k i/100ms
             Array#-    16.771k i/100ms
          two finger     6.942k i/100ms
        Set#include?    19.925k i/100ms
Calculating -------------------------------------
            include?    218.982k (± 1.9%) i/s -      1.103M in   5.038327s
       Regexp#match?      2.192M (± 2.2%) i/s -     11.077M in   5.056699s
       #match?(char)    184.918k (± 2.0%) i/s -    926.450k in   5.012209s
             Array#-    180.118k (± 1.6%) i/s -    905.634k in   5.029231s
          two finger     68.919k (± 1.4%) i/s -    347.100k in   5.037446s
        Set#include?    198.218k (± 1.6%) i/s -    996.250k in   5.027401s

Comparison:
       Regexp#match?:  2191627.2 i/s
            include?:   218981.8 i/s - 10.01x  (± 0.00) slower
        Set#include?:   198218.3 i/s - 11.06x  (± 0.00) slower
       #match?(char):   184918.5 i/s - 11.85x  (± 0.00) slower
             Array#-:   180118.1 i/s - 12.17x  (± 0.00) slower
          two finger:    68918.5 i/s - 31.80x  (± 0.00) slower
```

</details>

Unfortunately, it turns out that this isn't a hot path and leads to no performance gain. It's still useful to have though.

Relates to #23
Closes #27